pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gemini-imagegen"
version = "0.1.0"
description = "Generate images using Google's Gemini Nano Banana models"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = ["gemini", "image-generation", "ai", "google"]
dependencies = [
    "google-genai>=1.0.0",
    "Pillow>=10.0.0",
]

[project.scripts]
gemini-imagegen = "gemini_imagegen.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]
